hob parser: fix texture coord range

+++ hob_parser/hob_parser.pas
@@ -12,7 +12,7 @@ type
   end;
 
   TTexCoord = record
-      u, v: integer;
+      u, v: smallint;
   end;
 
   THobFace = record
